# docs/swagger.yaml
basePath: /
definitions:
  dto.SubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.SubscriptionResponse:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления подписками
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Возвращает список всех подписок с поддержкой пагинации
      parameters:
      - description: Максимальное количество элементов (по умолчанию 100)
        in: query
        name: limit
        type: integer
      - description: Смещение от начала (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список подписок
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionResponse'
            type: array
        "500":
          description: Ошибка при получении списка подписок
          schema:
            type: string
      summary: Список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку для пользователя
      parameters:
      - description: Данные новой подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная подписка
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            type: string
        "500":
          description: Ошибка при создании подписки
          schema:
            type: string
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет подписку по её ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Подписка успешно удалена
        "400":
          description: Некорректный ID
          schema:
            type: string
        "404":
          description: Подписка не найдена
          schema:
            type: string
        "500":
          description: Ошибка при удалении подписки
          schema:
            type: string
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Возвращает информацию о подписке по её ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Подписка
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Некорректный ID
          schema:
            type: string
        "404":
          description: Подписка не найдена
          schema:
            type: string
        "500":
          description: Ошибка при получении подписки
          schema:
            type: string
      summary: Получить подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет данные подписки по её ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Подписка успешно обновлена
        "400":
          description: Некорректные данные запроса
          schema:
            type: string
        "404":
          description: Подписка не найдена
          schema:
            type: string
        "500":
          description: Ошибка при обновлении подписки
          schema:
            type: string
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/{start}/{end}/total-cost:
    get:
      consumes:
      - application/json
      description: Считает суммарную стоимость подписок пользователя на конкретный
        сервис за период [start; end]
      parameters:
      - description: Начало периода (MM-YYYY)
        in: path
        name: start
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: path
        name: end
        required: true
        type: string
      - description: UUID пользователя (опционально)
        in: query
        name: user_id
        type: string
      - description: Название сервиса (опционально)
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Суммарная стоимость подписок
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Некорректные параметры запроса
          schema:
            type: string
        "500":
          description: Ошибка при вычислении стоимости
          schema:
            type: string
      summary: Общая стоимость подписок
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"